fix: Rename resourceDetailModule on store
