+++ src/store/index.js
@@ -5,12 +5,14 @@ import authReducer from "./modules/auth";
 import lectureReducer from "./modules/lecturesModule";
 import resourceReducer from "./modules/resourcesModule";
 import resourceCreateReducer from "./modules/resourceCreateModule";
+import resourceDetailReducer from "./modules/resourceDetailModule";
 
 const rootReducer = combineReducers({
   authReducer,
   lectureReducer,
   resourceReducer,
   resourceCreateReducer,
+  resourceDetailReducer,
 });
 
 const store = createStore(rootReducer, applyMiddleware(thunk));
