+++ src/store/modules/resourceDetailModule.js
